requirements.txt
# Web Framework
Flask==3.1.2
flask-cors==6.0.1
Werkzeug==3.1.3

# AI API
openai==2.7.1

# PDF Processing
PyPDF2==3.0.1

# Document Generation
python-docx==1.2.0

# Environment
python-dotenv==1.2.1

# HTTP client (for healthchecks)
requests==2.31.0

# Production WSGI server
gunicorn==23.0.0
